fix: require admin access for applicant routes and API
/applicants and /applicants/[profile] had no access control. Middleware
gated only /dashboard, /api and /admin, and neither page carried its own
session check — so both were statically prerendered and served as public,
cacheable HTML.

/api/applications was reachable by any signed-in account. Middleware only
tests `!user`, and the handler had no admin check of its own, so any member
could list every applicant record.

Today those surfaces only expose mock data, but they are the intended home
for real applications: names, emails, phone numbers, home addresses,
schools and essay answers.

- Add src/lib/auth/guards.ts with requireAdmin / getAdminUser /
  requireUser / getCurrentUser, so the check lives in one place instead of
  being re-derived per route.
- Convert both applicant routes to server components that call
  requireAdmin(), moving the existing client code to ApplicantsPageClient
  and ApplicantProfileClient. This mirrors the pattern admin/page.tsx
  already used.
- Return 403 from /api/applications for non-admins.
- Add /applicants to USER_PATHS as a first gate. Note middleware alone is
  not sufficient — it cannot distinguish admins from members.
- Refactor admin/page.tsx onto the shared guard, removing its duplicate.

`next build` now reports /applicants and /applicants/[profile] as dynamic
(f) rather than static (o), confirming the guard runs.

Also aligns this file's Supabase key name with the rest of the codebase;
see the following commit for that change.
