+++ client/src/app/admin/page.tsx
@@ -1,28 +1,9 @@
-import { redirect } from "next/navigation";
 import AdminPageClient from "./AdminPageClient";
 import AddCalendly from "./AddCalendly";
-import { createClient } from "@/lib/supabase/server";
+import { requireAdmin } from "@/lib/auth/guards";
 
 export default async function AdminPage() {
-  const supabase = await createClient();
-
-  const {
-    data: { user },
-  } = await supabase.auth.getUser();
-
-  if (!user) {
-    return redirect("/auth");
-  }
-
-  const { data: userRow, error } = await supabase
-    .from("user")
-    .select("*")
-    .eq("id", user.id)
-    .single();
-
-  if (error || !userRow?.admin) {
-    return redirect("/auth");
-  }
+  const userRow = await requireAdmin();
 
   // pass user data to client if needed (has to be plain data)
   return (
